# core/docs/swagger.yaml
definitions:
  internal_api_handlers.AddressCreateRequest:
    properties:
      ips:
        items:
          type: string
        type: array
      name:
        type: string
    type: object
  internal_api_handlers.AddressEditRequest:
    properties:
      ips:
        items:
          type: string
        type: array
    type: object
  internal_api_handlers.DhcpCreateRequest:
    properties:
      dns:
        items:
          type: string
        type: array
      end_ip:
        type: string
      gateway:
        type: string
      interface:
        type: string
      lease_time:
        type: integer
      start_ip:
        type: string
    type: object
  internal_api_handlers.DhcpEditRequest:
    properties:
      dns:
        items:
          type: string
        type: array
      end_ip:
        type: string
      gateway:
        type: string
      lease_time:
        type: integer
      start_ip:
        type: string
    type: object
...
  net.IPNet:
    properties:
      ip:
        description: network number
        items:
          format: int32
          type: integer
        type: array
      mask:
        description: network mask
        items:
          format: int32
          type: integer
        type: array
    type: object
...
  sentinelos_core_internal_model.Address:
    properties:
      ips:
        items:
          $ref: '#/definitions/net.IPNet'
        type: array
      name:
        type: string
    type: object
  sentinelos_core_internal_model.DHCP:
    properties:
      dns:
        items:
          type: string
        type: array
      endIP:
        type: string
      gateway:
        type: string
      interface:
        type: string
      leaseTimeMin:
        type: integer
      startIP:
        type: string
      subnetMask:
        type: string
    type: object
...
  sentinelos_core_pkg_utils.APIError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
  description: Next-Generation Firewall Management API.
  title: SentinelOS API
  version: "1.0"
paths:
  /api/addresses:
    get:
      description: Get a list of all addresses.
      produces:
      - application/json
      responses:
        "200":
          description: List of addresses
          schema:
            additionalProperties:
              $ref: '#/definitions/sentinelos_core_internal_model.Address'
            type: object
        "500":
          description: ERR_SYS_4001 Internal server error
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: List Addresses
      tags:
      - networking
      - addresses
    post:
      consumes:
      - application/json
      description: Create a new address object.
      parameters:
      - description: Address details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.AddressCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: address created in candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ERR_NET_1004 Invalid JSON or Missing field
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
        "409":
          description: ERR_NET_2006 Resource already exists
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Create Address
      tags:
      - networking
      - addresses
  /api/addresses/{name}:
    delete:
      description: Delete an existing address object.
      parameters:
      - description: Address Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: address deleted from candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Missing required field
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
        "404":
          description: ERR_NET_1003 Resource not found
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Delete Address
      tags:
      - networking
      - addresses
    put:
      consumes:
      - application/json
      description: Update an existing address object.
      parameters:
      - description: Address Name
        in: path
        name: name
        required: true
        type: string
      - description: Address details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.AddressEditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: address updated in candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid JSON or Missing field
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
        "404":
          description: ERR_NET_1003 Resource not found
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Edit Address
      tags:
      - networking
      - addresses
  /api/config/begin:
    post:
      description: Start a new candidate configuration session.
      produces:
      - application/json
      responses:
        "200":
          description: 'message: config session started'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Missing authorization token
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
        "409":
          description: Config session already active
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Begin Configuration Session
      tags:
      - system
      - config
  /api/config/commit:
    post:
      description: Commit the candidate configuration to running configuration.
      produces:
      - application/json
      responses:
        "200":
          description: 'message: commit successful'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Commit error
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Commit Configuration
      tags:
      - system
      - config
  /api/dhcp:
    get:
      description: Get a list of all DHCP configurations.
      produces:
      - application/json
      responses:
        "200":
          description: List of DHCP configs
          schema:
            items:
              $ref: '#/definitions/sentinelos_core_internal_model.DHCP'
            type: array
        "500":
          description: ERR_SYS_4001 Internal server error
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: List DHCP Configs
      tags:
      - networking
      - dhcp
    post:
      consumes:
      - application/json
      description: Create a new DHCP configuration.
      parameters:
      - description: DHCP details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.DhcpCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: dhcp pool created in candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Create DHCP Config
      tags:
      - networking
      - dhcp
  /api/dhcp/{interface}:
    delete:
      description: Delete an existing DHCP configuration.
      parameters:
      - description: Interface Name
        in: path
        name: interface
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: dhcp pool deleted in candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Missing required field
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
        "404":
          description: ERR_NET_1003 Resource not found
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Delete DHCP Config
      tags:
      - networking
      - dhcp
    put:
      consumes:
      - application/json
      description: Update an existing DHCP configuration.
      parameters:
      - description: Interface Name
        in: path
        name: interface
        required: true
        type: string
      - description: DHCP details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.DhcpEditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: dhcp pool updated in candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid JSON or Missing field
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
        "404":
          description: ERR_NET_1003 Resource not found
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Edit DHCP Config
      tags:
      - networking
      - dhcp
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"